Add a mode switch to support EXPANDING mode as well as CONSTRAINED
Constrained is the old mode, where all panes are constrained by the size of their parent container, and each drag just partitions the space between two adjacent panes. In Expanding mode, the size of the pane to the left of the dragged gutter expands or contracts like before, but the parent container expands and contracts along with it instead of the pane to the right of the gutter changing its size.